feat: added association of products controllers and database table with their users.

+++ src/controllers/products.ts
@@ -44,13 +44,15 @@ export default class ProductsController {
 		}
 	}
 
-	async createProduct(req: Request, res: Response) {
+	async createProduct(req: Request, res: Response): ControllerResponse {
+		const new_product = req.body;
+		const user_id = (req as any).user.id;
+		new_product.createdBy = user_id;
+		if (!new_product)
+			return res
+				.status(400)
+				.json({ message: 'No data received to save a new product.' });
 		try {
-			const new_product = req.body;
-			if (!new_product) {
-				res.status(400).json({ message: 'No data received.' });
-				return;
-			}
 			await Product.create({ ...new_product }, { returning: false });
 			res.status(201).json({ message: 'Product created successfuly.' });
 		} catch (err) {
@@ -58,22 +60,22 @@ export default class ProductsController {
 		}
 	}
 
-	async updateProduct(req: Request, res: Response) {
+	async updateProduct(req: Request, res: Response): ControllerResponse {
+		const product_id = Number(req.params.id);
+		const user_id = (req as any).user.id;
+		const updatedProduct = req.body;
+		if (!product_id)
+			return res
+				.status(400)
+				.json({ message: 'Provided product ID is invalid.' });
+		if (!updatedProduct)
+			return res
+				.status(400)
+				.json({ message: 'No data received for this update operation.' });
 		try {
-			const product_id = Number(req.params.id);
-			const updatedProduct = req.body;
-			if (!product_id) {
-				res.status(400).json({ message: 'Provided product ID is invalid.' });
-				return;
-			} else if (!updatedProduct) {
-				res
-					.status(400)
-					.json({ message: 'No data received for this update operation.' });
-				return;
-			}
 			await Product.update(
 				{ ...updatedProduct },
-				{ where: { id: product_id }, returning: false }
+				{ where: { id: product_id, createdBy: user_id }, returning: false }
 			);
 			res.status(200).json({ message: 'Product updated successfuly.' });
 		} catch (err) {
@@ -81,14 +83,15 @@ export default class ProductsController {
 		}
 	}
 
-	async deleteProduct(req: Request, res: Response) {
+	async deleteProduct(req: Request, res: Response): ControllerResponse {
+		const product_id = Number(req.params.id);
+		const user_id = (req as any).user.id;
+		if (!product_id)
+			return res
+				.status(400)
+				.json({ message: 'Provided product ID is invalid.' });
 		try {
-			const product_id = Number(req.params.id);
-			if (!product_id) {
-				res.status(400).json({ message: 'Provided product ID is invalid.' });
-				return;
-			}
-			await Product.destroy({ where: { id: product_id } });
+			await Product.destroy({ where: { id: product_id, createdBy: user_id } });
 			res.status(200).json({ message: 'Product data deleted successfuly.' });
 		} catch (err) {
 			res.status(500).json({ err });
